added obd2 unit tests

+++ Software/OBD2/inc/obd2.hpp
@@ -21,8 +21,8 @@ class obd2{
         float decoded_cmd(string cmd);
         vector<string> read_dtc();
         string dtc_desc(string dtc_code);
-        int hex2int(string hex_string);
-        string int2hex(int int_value);
+        unsigned long hex2int(string hex_string);
+        string int2hex(long int_value);
     
     private:
